+++ app/(landing)/signup/information/page.tsx
@@ -1,8 +1,6 @@
 'use client'
 
-import { ChangeEvent, useState } from 'react'
-
-import { useRouter } from 'next/navigation'
+import { ChangeEvent } from 'react'
 
 import classNames from 'classnames/bind'
 
@@ -13,38 +11,13 @@ import { Input } from '@/shared/ui/input'
 import { LinkButton } from '@/shared/ui/link-button'
 import Select from '@/shared/ui/select'
 
+import useSignupEmail from '../_hooks/custom/use-signup-email'
+import useSignupForm from '../_hooks/custom/use-signup-form'
 import Step from '../_ui/step'
 import styles from './page.module.scss'
-import { SignupFormDataModel, SignupFormErrorsModel, SignupFormStateModel } from './types'
-import {
-  generateDayOptions,
-  generateMonthOptions,
-  generateYearOptions,
-  validateSignupForm,
-} from './utils'
+import { generateDayOptions, generateMonthOptions, generateYearOptions } from './utils'
 
 const cx = classNames.bind(styles)
-
-const initialForm: SignupFormDataModel = {
-  name: '',
-  nickname: '',
-  email: '',
-  emailDomain: '',
-  verificationCode: '',
-  password: '',
-  passwordConfirm: '',
-  phone: '',
-  birthYear: '',
-  birthMonth: '',
-  birthDay: '',
-  isMarketingAgreed: false,
-}
-
-const initialFormState: SignupFormStateModel = {
-  isEmailVerified: false,
-  isNicknameVerified: false,
-  isEmailSent: false,
-}
 
 const emailDomainOptions = [
   { value: '', label: '직접 입력' },
@@ -63,116 +36,41 @@ const selectStyle = {
 }
 
 const InformationPage = () => {
-  const router = useRouter()
-  const [form, setForm] = useState<SignupFormDataModel>(initialForm)
-  const [errors, setErrors] = useState<SignupFormErrorsModel>({})
-  const [formState, setFormState] = useState<SignupFormStateModel>(initialFormState)
-  const [isValidated, setIsValidated] = useState(false)
-  const [selectedDomain, setSelectedDomain] = useState<string>('')
+  const {
+    form,
+    setForm,
+    errors,
+    setErrors,
+    formState,
+    setFormState,
+    isValidated,
+    handleInputChange,
+    handleMarketingAgree,
+    handleFormSubmit,
+    handleNicknameCheck,
+  } = useSignupForm()
 
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target
-    setForm((prev) => ({ ...prev, [name]: value }))
+  const {
+    selectedDomain,
+    handleEmailChange,
+    handleDomainSelect,
+    handleEmailVerification,
+    handleVerificationCodeCheck,
+  } = useSignupEmail({
+    errors,
+    isValidated,
+    setForm,
+    setErrors,
+    setFormState,
+  })
 
-    if (isValidated) {
-      setErrors((prev) => ({
-        ...prev,
-        [name]: null,
-      }))
-    }
-
-    if (name === 'emailDomain') {
-      setSelectedDomain('')
-    }
+  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const { name } = e.target
+    handleInputChange(e)
 
     if (name === 'email' || name === 'emailDomain') {
-      setFormState((prev) => ({
-        ...prev,
-        isEmailVerified: false,
-        isEmailSent: false,
-      }))
+      handleEmailChange(name)
     }
-
-    if (name === 'nickname') {
-      setFormState((prev) => ({
-        ...prev,
-        isNicknameVerified: false,
-      }))
-    }
-  }
-
-  const handleDomainSelect = (value: string | string[]) => {
-    if (typeof value === 'string') {
-      setSelectedDomain(value)
-      setForm((prev) => ({ ...prev, emailDomain: value !== '' ? value : '' }))
-
-      if (isValidated) {
-        setErrors((prev) => ({ ...prev, email: null }))
-      }
-
-      setFormState((prev) => ({
-        ...prev,
-        isEmailVerified: false,
-        isEmailSent: false,
-      }))
-    }
-  }
-
-  const handleNicknameCheck = async () => {
-    try {
-      // TODO: 닉네임 중복 확인 API 호출
-      setFormState((prev) => ({ ...prev, isNicknameVerified: true }))
-
-      if (errors.nickname) {
-        setErrors((prev) => ({ ...prev, nickname: null }))
-      }
-    } catch (error) {
-      console.error('닉네임 중복 확인 실패:', error)
-    }
-  }
-
-  const handleEmailVerification = async () => {
-    try {
-      // TODO: 이메일 인증 API 호출
-      setFormState((prev) => ({ ...prev, isEmailSent: true }))
-    } catch (error) {
-      console.error('이메일 인증 발송 실패:', error)
-    }
-  }
-
-  const handleVerificationCodeCheck = async () => {
-    try {
-      // TODO: 인증번호 확인 API 호출
-      setFormState((prev) => ({ ...prev, isEmailVerified: true }))
-
-      if (errors.email) {
-        setErrors((prev) => ({ ...prev, email: null }))
-      }
-    } catch (error) {
-      console.error('인증번호 확인 실패:', error)
-    }
-  }
-
-  const handleMarketingAgree = (checked: boolean) => {
-    setForm((prev) => ({ ...prev, isMarketingAgreed: checked }))
-  }
-
-  const handleFormSubmit = () => {
-    const formErrors = validateSignupForm(
-      form,
-      formState.isEmailVerified,
-      formState.isNicknameVerified
-    )
-    setIsValidated(true)
-
-    if (Object.keys(formErrors).length > 0) {
-      setErrors(formErrors)
-      return
-    }
-
-    // TODO: 회원가입 API 호출
-    console.log('폼 제출', form)
-    router.push(PATH.SIGN_UP_COMPLETE)
   }
 
   return (
@@ -223,7 +121,7 @@ const InformationPage = () => {
                   id="email"
                   name="email"
                   value={form.email}
-                  onChange={handleInputChange}
+                  onChange={handleChange}
                   placeholder="이메일 주소를 입력하세요"
                   className={cx('input')}
                   errorMessage={errors.email}
@@ -233,7 +131,7 @@ const InformationPage = () => {
                   id="emailDomain"
                   name="emailDomain"
                   value={form.emailDomain}
-                  onChange={handleInputChange}
+                  onChange={handleChange}
                   inputSize="small"
                   className={cx('input')}
                   disabled={selectedDomain !== ''}
